requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
dir-tree==0.0.2
Django==5.2.1
et_xmlfile==2.0.0
frozenlist==1.6.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
openpyxl==3.1.5
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
setuptools==80.9.0
sqlparse==0.5.3
svgwrite==1.4.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
yarl==1.20.0
gunicorn
psycopg2-binary
